2022-3-23 by huyuan 1） go-cms 框架升级到beego 2.0.2 2） gorm暂时未动，v1 到v2 跨度太大。 3） 数据库在整理中，已发现模型不同步的问题。 4） 记录一些整理的笔记。

+++ pkg/d/api_return.go
@@ -7,7 +7,7 @@ import (
 )
 
 //普通json格式
-func ReturnJson(code int, msg string, data interface{}) (jsonData map[string]interface{}) {
+func ReturnJson(code int64, msg string, data interface{}) (jsonData map[string]interface{}) {
 	jsonData = make(map[string]interface{}, 3)
 	jsonData["time_stamp"] = time.Now()
 	jsonData["code"] = code
@@ -29,12 +29,12 @@ func ReturnParamErrJson(data interface{}) (map[string]interface{}) {
 }
 
 //layui 后台返回需要的json格式
-func LayuiJson(code int, msg string, data, count , pageNo , pageSize  interface{}) (jsonData map[string]interface{}) {
+func LayuiJson(code int64, msg string, data, count , pageNo , pageSize  interface{}) (jsonData map[string]interface{}) {
 	jsonData = make(map[string]interface{}, 3)
 	jsonData["code"] = code
 	jsonData["msg"] = msg
 	if count !=false {
-		jsonData["data"] = util.PageUtil(count.(int64) , pageNo.(int64) , pageSize.(int64) , data)
+		jsonData["data"] = util.PageUtil(count.(int64) , pageNo.(int) , pageSize.(int) , data)
 	}else {
 		jsonData["data"] = data
 	}
